# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Swagger hujjatlari localhost serveri uchun",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:99",
            "description": "Localhost Server"
        }
    ],
    "paths": {
        "/api/operators": {
            "get": {
                "tags": [
                    "Operators"
                ],
                "summary": "Foydalanuvchilar ro'yxati",
                "operationId": "8a2d53adab8cb62cf3c9a053901c6e9f",
                "responses": {
                    "200": {
                        "description": "Muvaffaqiyatli",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Operator"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Operators"
                ],
                "summary": "Create a new operator",
                "operationId": "26809c6865c50e48b20c79e8737f2763",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "chat_id",
                                    "is_supervisor"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "chat_id": {
                                        "type": "integer",
                                        "example": "your telegram chat_id"
                                    },
                                    "is_supervisor": {
                                        "description": "Agar admin bo‘lsa true, oddiy operator bo‘lsa false",
                                        "type": "boolean",
                                        "example": "false"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Author created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/operators/{id}": {
            "get": {
                "tags": [
                    "Operators"
                ],
                "summary": "Get operator by ID",
                "operationId": "a4c4c136b5b1beb0748577008ec5e6c8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operator",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Operator"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Operator not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Operators"
                ],
                "summary": "Update an operator by ID",
                "operationId": "d129f42821fdf9d80fd68f94e6eb6da3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the operator to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Yangi Nomi"
                                    },
                                    "chat_id": {
                                        "type": "integer",
                                        "example": 123456789
                                    },
                                    "is_supervisor": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Operator updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Operator"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Operator not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Operators"
                ],
                "summary": "Delete an operator by ID",
                "operationId": "39e89f11a9a6d53d5d1815cd6b63afd7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the operator to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Operator deleted successfully"
                    },
                    "404": {
                        "description": "Operator not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Operator": {
                "title": "Operator",
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Operators",
            "description": "Operators"
        }
    ]
}